BeforeDelete added to Story, All tests pass
Getting foreign key issues again with the editor collection on Story.
Had worked before but maybe since Hibernate upgrade.  Now not doing the
cascade delete so created beforeDelete on Story, added removeAll on
Editor.  Couldn't use a where {} clause to delete editors as it
complained of a not being able to DELETE or UPDATE on a JOIN.  The query
did not do a JOIN.  Changed to findAllByStory(s)*.delete()